backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
google-genai>=0.3.0
python-multipart>=0.0.9
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
